// txmatching/web/frontend/src/app/generated/model/antibodiesPerGroupGenerated.ts
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HlaAntibodyGenerated } from './hlaAntibodyGenerated';


export interface AntibodiesPerGroupGenerated {
    hla_antibody_list: Array<HlaAntibodyGenerated>;
    hla_group: AntibodiesPerGroupGeneratedHlaGroupEnum;
}
export enum AntibodiesPerGroupGeneratedHlaGroupEnum {
    A = 'A',
    B = 'B',
    Drb1 = 'DRB1',
    Other = 'Other'
};