// txmatching/web/frontend/src/app/generated/model/antigenMatchGenerated.ts
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HlaTypeGenerated } from './hlaTypeGenerated';


export interface AntigenMatchGenerated {
    hla_type: HlaTypeGenerated;
    match_type: AntigenMatchGeneratedMatchTypeEnum;
}
export enum AntigenMatchGeneratedMatchTypeEnum {
    Split = 'SPLIT',
    Broad = 'BROAD',
    HighRes = 'HIGH_RES',
    None = 'NONE'
};